@@ -9,12 +9,14 @@
     "axios": "^0.19.0",
     "history": "^4.3.0",
     "marked": "^0.3.6",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
-    "react-redux": "^4.4.5",
-    "react-router": "^2.8.1",
-    "redux": "^3.6.0",
+    "react": "^16.9.0",
+    "react-dom": "^16.9.0",
+    "react-redux": "^7.0.2",
+    "react-router-dom": "^5.0.0",
+    "redux": "^4.0.1",
+    "redux-devtools-extension": "^2.13.8",
     "redux-thunk": "^2.3.0",
+    "redux-saga": "^1.1.1",
     "superagent": "^2.3.0",
     "superagent-promise": "^1.1.0"
   },
