@@ -49,7 +49,7 @@ export const CREATE_COMMENT = "CREARE_COMMENT";
 export const DEL_COMMENT = "DEL_COMMENT";
 
 //READ
-export const GET_COMMENT_FROM_ARTICLE = "GET_COMMENT_FROM_ARTICLE"
+export const GET_COMMENTS = "GET_COMMENTS"
 
 
 ///////////
@@ -66,28 +66,15 @@ export const UNFOL_USER = "UNFOL_USER";
 export const GET_USER = "GET_USER"
 
 
-// export const SETTINGS_SAVED = "SETTINGS_SAVED";
-// export const DELETE_ARTICLE = "DELETE_ARTICLE";
-// export const SETTINGS_PAGE_UNLOADED = "SETTINGS_PAGE_UNLOADED";
-// export const HOME_PAGE_LOADED = "HOME_PAGE_LOADED";
 export const HOME_PAGE_UNLOADED = "HOME_PAGE_UNLOADED";
 
-// export const ARTICLE_PAGE_LOADED = "ARTICLE_PAGE_LOADED";
-// export const ARTICLE_PAGE_UNLOADED = "ARTICLE_PAGE_UNLOADED";
-
-// export const ADD_COMMENT = "ADD_COMMENT";
-// export const DELETE_COMMENT = "DELETE_COMMENT";
-// export const ARTICLE_FAVORITED = "ARTICLE_FAVORITED";
-// export const ARTICLE_UNFAVORITED = "ARTICLE_UNFAVORITED";
-// export const SET_PAGE = "SET_PAGE";
-// export const APPLY_TAG_FILTER = "APPLY_TAG_FILTER";
-// export const CHANGE_TAB = "CHANGE_TAB";
-
-// export const PROFILE_PAGE_LOADED = "PROFILE_PAGE_LOADED";
 
 //AUTH
 
 export const LOGIN = "LOGIN";
+//saga
+export const LOGIN_SUCCESS = "LOGIN_SUCCESS"
+
 export const LOGOUT = "LOGOUT";
 export const REGISTER = "REGISTER";
 export const LOGIN_PAGE_UNLOADED = "LOGIN_PAGE_UNLOADED";
@@ -102,15 +89,8 @@ export const USER_LOADING = "USER_LOADING";
 // export const ADD_TAG = "ADD_TAG";
 // export const REMOVE_TAG = "REMOVE_TAG";
 
-// Kind of unnecessary, because I can use state instead of using redux to handle onchange event.
-
-// export const UPDATE_FIELD_AUTH = 'UPDATE_FIELD_AUTH';
-// export const UPDATE_FIELD_EDITOR = 'UPDATE_FIELD_EDITOR';
-
-//FOLLOW and UNFOLLOW USER
-// export const FOLLOW_USER = "FOLLOW_USER";
-// export const UNFOLLOW_USER = "UNFOLLOW_USER";
-
 export const PROFILE_FAVORITES_PAGE_UNLOADED =
   "PROFILE_FAVORITES_PAGE_UNLOADED";
 // export const PROFILE_FAVORITES_PAGE_LOADED = "PROFILE_FAVORITES_PAGE_LOADED";
+
+export const AUTH_SUCCESS = "AUTH_SUCCESS"
