@@ -1,7 +1,6 @@
-import * as GET from "../action/type";
+import * as actionType from "../action/type";
 
 const initialState = {
-  token: localStorage.getItem("token"),
   isAuth: null,
   loading: true,
   user: null
@@ -11,29 +10,20 @@ export default (state = initialState, action) => {
   const { type, payload } = action;
 
   switch (type) {
-    case GET.USER_LOADED:
-      return {
-        ...state,
-        isAuth: true,
-        loading: false,
-        user: payload
-      };
-    case GET.LOGIN:
-    case GET.REGISTER:
-      localStorage.setItem("token", payload.user.token);
+   
+    case actionType.AUTH_SUCCESS:
       return {
         ...state,
         loading: false,
         isAuth: true,
-        user: payload
+        user: payload.user
       };
 
-    case GET.LOGIN_PAGE_UNLOADED:
-    case GET.REGISTER_PAGE_UNLOADED:
+    case actionType.LOGIN_PAGE_UNLOADED:
+    case actionType.REGISTER_PAGE_UNLOADED:
       return {};
 
-    case GET.LOGOUT: 
-      localStorage.removeItem("token");
+    case actionType.LOGOUT: 
       return {
         ...state,
         token: null,
